Improve candidate list comments

+++ Sparse_CCJ/base.h
@@ -5,10 +5,9 @@
 #include <utility>
 
 //! type of energy
-/// TODO short?
-typedef int energy_t;
-/// TODO could be unsigned but probably unnecessary
-typedef short index_t;
+// Should not be a short as INF is 1600000 and is a very common value for energy.
+typedef long int energy_t;
+typedef unsigned short index_t;
 
 
 template<class T1,class T2>
